remote post (see commit mesg. desc)
TL;DR: This was a terrible piece that shouldn't have gone up. It may return after significant revisions, but it was just a generally not great post.

---

I enjoy writing these posts, but sometimes I feel the urge to write even if it's a topic no one cares about or I'm writing absolute non-sense. This post about TLD reputability isn't *inaccurate*, it's just poorly constructed. What I want to do moving forward is craft posts of a quality that would rival that of a thesis while being readable by a junior high student. It allows that nice balance between accessibility with new people being able to understand more complex concepts, as well as give me that tiny "yeah, I'm teaching people stuff!"

In short, this post is going to be taken down for the foreseeable future and remain (clearly) visible in commit history. I may revise this post and release it again in the future, but this is the best course of the time being.